feat: 차량 삭제 시, 없는 id일 경우 exception 발생 시킴

+++ admin/src/main/java/kernel360/ckt/admin/application/VehicleService.java
@@ -49,7 +49,15 @@ public class VehicleService {
             .orElseThrow(() -> new EntityNotFoundException("Vehicle not found with id: " + vehicleId));
     }
 
+//    public void delete(Long vehicleId) {
+//        vehicleRepository.deleteById(vehicleId)
+//            .orElseThrow(() -> new EntityNotFoundException("Vehicle not found with id: " + vehicleId);
+//    }
+
     public void delete(Long vehicleId) {
+        if (vehicleRepository.findById(vehicleId).isEmpty()) {
+            throw new EntityNotFoundException("Vehicle not found with id: " + vehicleId);
+        }
         vehicleRepository.deleteById(vehicleId);
     }
 }
